Fix Build warnings to do with BufferElement Offset Type

+++ Crosa-Conty-3D/src/CC3D/Renderer/Buffer.h
@@ -32,7 +32,7 @@ namespace CC3D {
 		std::string Name;
 		ShaderDataType Type;
 		uint32_t Size;
-		uint32_t Offset;
+		size_t Offset;
 		bool Normalized;
 
 		BufferElement() = default;
@@ -85,7 +85,7 @@ namespace CC3D {
 	private:
 		void CalculateOffsetsAndStride()
 		{
-			uint32_t offset = 0;
+			size_t offset = 0;
 			m_Stride = 0;
 			for (auto& element : m_Elements)
 			{
